Update README.md
Overview of Project ☁️
In this project, we will be building an image labels generator using Amazon Rekognition. This is going to be a fun one. Once built, it will be able to recognise and label images. For example, if you have a photo of a cat, Amazon Rekognition will be able to identify it and label the image as a cat.